initial commit
Is-your-birthday-palindrome
This is a fun web app that takes your birthday and tells you if your birthday is a palindrome. In case your bithday is not a palindrome it tells you the next palindrome date from the input date.

This project is built using Vanilla JS, HTML and CSS.

The site has two javascript file that implements the same output but are the two different programs. The main.js is the main file linked with the website. The alternative.js is an alternatie file that execute similar output.

The program checks if the input date is a palindrome, across various date formats given below: DD-MM-YYYY MM-DD-YYYY YYYY-MM-DD DD-MM-YY MM-DD-YY YY-MM-DD

I build this project as a part of a slew of projects in Neog.camp level zero curriculum. The program offers minimum hand holding and the available materials are enough to help whenever need arises.

Please visit my portfolio site for other such projects, I'll really appreciate and thank you for your time.